Add extension in file context menu
added required column Extension, which can be pipe "|" separated list of extensions. ie .txt|.png|.md

Adds to current fileType, or creates a new file type if necessary. Does not override user choice, must manually do it from filetypesman.

Remarks - Only one or the other is supported (global ext's menu or specific ext menu). If you need to add it to both for some reason, enter them as separate lines in the csv

Aux
RegWrite Wrapper function for native command. Includes params "overwrite" and "createNew". Since RegRead is unable to determine if a key's value is not existent vs just being empty string, we can pass cfg that specifies if it should be overwritten, or created if not there.

TODO
replace regwrite commands with the wrapper