allow pipe delimited ext for default programs
allow pipe delimited ext for default programs

allow multiple file ext per default program in csv as a convenience. separate filetypes will still be created for each ext, even if they are associated with the same program.

The goal is to unlink user choice, define default program for multiple ext's, but still have the ability to define custom actions on  specific file types. With user choice and shared filetypes, you are forced to add those context menu items for all the associated ext's, which may not always be desired.