dont inject default param if command is running a start command from cmd
in some cases we may need to start a program through cmd. For powershell_ise, we want to start it in a directory. However, its default param tries to open the passed param as a ps1 file in the editor view, instead of opening the passed value as the directory in the shell view.

Instead, we can run cmd (which by default runs in the path we triggered the context menu) and then start powershell_ise, which when started in the directory, it will default to that directory in the bottom shell.

we inject a default parameter in most cases, but dont do it here as it causes issue.