support file type in csv
User can pass FileType in csv for when they want to specify a filetype. Only applicable when extensions is populated.

Most of the time we want to create our own file types, but for some system default ones like powershell, it is more convenient to use the existing ones.

Works for both default programs and specific file type context menus

AUX
-default programs command needs to use expandable string for reg value type